Add the ability to override res, req creation
so that a customized HTTPServer subclass can use it's own Request/Response classes.

To apply the override, make a subclass of WEBrick::HTTPServer and override the
`create_request_and_response(with_webrick_config)` method. The method should return
an Array of [request, response].

To check whether the Server supports this method (i.e. when using older versions of
WEBrick when needing this functionality), you can ask the server if it responds
to the method

    server.respond_to?(:create_request_and_response)

This is backportable.